+++ libft/ft_memcpy.c
@@ -6,11 +6,14 @@
 /*   By: pdovalo- <pdovalo-@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/20 13:24:23 by pdovalo-          #+#    #+#             */
-/*   Updated: 2024/09/20 13:36:47 by pdovalo-         ###   ########.fr       */
+/*   Updated: 2024/09/25 14:12:44 by pdovalo-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dest, const void *src, unsigned int n)
+#include <string.h>
+#include <libft.h>
+
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
 	unsigned int	i;
 	char			*dest_a;
