+++ libft/ft_memmove.c
@@ -6,13 +6,15 @@
 /*   By: pdovalo- <pdovalo-@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/20 13:57:37 by pdovalo-          #+#    #+#             */
-/*   Updated: 2024/09/25 14:01:27 by pdovalo-         ###   ########.fr       */
+/*   Updated: 2024/09/25 14:13:25 by pdovalo-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 //#include <stdio.h>
+#include <string.h>
+#include <libft.h>
 
-void	*ft_memmove(void *dest, const void *src, unsigned int n)
+void	*ft_memmove(void *dest, const void *src, size_t n)
 {
 	unsigned int	i;
 	char			*dest_a;
