+++ libft/ft_memset.c
@@ -6,11 +6,14 @@
 /*   By: pdovalo- <pdovalo-@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/20 12:29:24 by pdovalo-          #+#    #+#             */
-/*   Updated: 2024/09/20 13:17:45 by pdovalo-         ###   ########.fr       */
+/*   Updated: 2024/09/25 14:14:20 by pdovalo-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *str, int c, unsigned int n)
+#include <string.h>
+#include <libft.h>
+
+void	*ft_memset(void *str, int c, size_t n)
 {
 	unsigned int	i;
 	char			*str_a;
